requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
python-multipart
fastapi-cors
python-jose[cryptography]
passlib[bcrypt]
google-genai